switch to working with doubles instead of float

+++ src/main/java/components/Player.java
@@ -2,6 +2,7 @@ package components;
 
 import engine.Camera;
 import engine.Window;
+import org.joml.Vector2d;
 import org.joml.Vector2f;
 import scenes.Scene;
 
@@ -12,7 +13,7 @@ public class Player extends Component{
         Scene scene = Window.getInstance().getScene();
         Camera camera = scene.getCamera();
         scene.setPlayer(this.gameObject);
-        camera.setViewPoint(new Vector2f(this.gameObject.getX() - camera.getProjectionSize().x  / 2,
+        camera.setViewPoint(new Vector2d(this.gameObject.getX() - camera.getProjectionSize().x  / 2,
                 this.gameObject.getY() - camera.getProjectionSize().y / 2));
     }
 }
