+++ src/main/java/renderer/Transform.java
@@ -1,32 +1,32 @@
 package renderer;
 
-import org.joml.Vector2f;
+import org.joml.Vector2d;
 
 public class Transform {
 
-    public Vector2f position;
-    public Vector2f scale;
+    public Vector2d position;
+    public Vector2d scale;
 
 
     public Transform(){
-        init(new Vector2f(), new Vector2f());
+        init(new Vector2d(), new Vector2d());
     }
 
     public Transform(Transform tr){
-        init(new Vector2f(tr.position), new Vector2f(tr.scale));
+        init(new Vector2d(tr.position), new Vector2d(tr.scale));
     }
 
-    public Transform(Vector2f position, Vector2f scale) {
+    public Transform(Vector2d position, Vector2d scale) {
         init(position, scale);
     }
 
-    public void init(Vector2f position, Vector2f scale) {
+    public void init(Vector2d position, Vector2d scale) {
         this.position = position;
         this.scale = scale;
     }
 
     public Transform copy() {
-        return new Transform(new Vector2f(this.position), new Vector2f(this.scale));
+        return new Transform(new Vector2d(this.position), new Vector2d(this.scale));
     }
 
     public void copy(Transform to) {
@@ -38,11 +38,11 @@ public class Transform {
         return (this.scale.equals(transform.scale) && this.position.equals(transform.position));
     }
 
-    public void setPosition(Vector2f position) {
+    public void setPosition(Vector2d position) {
         this.position = position;
     }
 
-    public void setScale(Vector2f scale) {
+    public void setScale(Vector2d scale) {
         this.scale = scale;
     }
 }
